+++ tests/testthat/test-lidR-compatibility.R
@@ -36,20 +36,3 @@ test_that("lasR supports a LAScatalog from lidR",
 
   expect_error(exec(pipeline, on = las), NA)
 })
-
-test_that("exec works with old pipeline",
-{
-  f <- system.file("extdata", "Megaplot.las", package="lasR")
-
-  r = reader(f, filter = drop_z_below(2)) |> suppressWarnings()
-  tri = triangulate(25)
-  h = hulls(tri)
-  pipeline = r  + tri + h
-
-  ans1 = exec(pipeline) |> suppressWarnings()
-  ans2 = exec(pipeline, on = f) |> suppressWarnings()
-
-  expect_s3_class(ans1, "sf")
-  expect_equal(nrow(ans1$geom[[1]][[1]]), 85L)
-  expect_equal(nrow(ans2$geom[[1]][[1]]), 85L)
-})